requirements.txt
Django
django-cors-headers
django-filter
django-channels-graphql-ws==0.9.1
graphene-django
channels[daphne]
